fixed end-point bug

+++ endpoints/sell.py
@@ -3,7 +3,7 @@ from flask import request, g, abort
 from flask_restful import Resource
 from extensions import auth, db
 from endpoints import api
-from models import UserLog, UserItem
+from models import UserLog, UserItem, Item
 
 logger = logging.getLogger(__name__)
 
@@ -18,18 +18,21 @@ class Sell(Resource):
         user_item = UserItem.query.filter_by(user_id=g.current_user.id, item_id=item_id).first()
         if user_item is None: abort(400, f"You don't own any items with id [{item_id}]") 
 
+        item = Item.query.filter_by(id=item_id).first()
+        if item is None: abort(400, f"This is for an item which doesn't exist... [{item_id}]") 
+        
         sell_amount = max(int(request.json.get("amount", 1)), 1)
         if sell_amount > user_item.amount: abort(400, f"You don't have [{sell_amount}] of item [{user_item.item.name}] to sell. You have [{user_item.amount}] in stock.")
 
         transaction_fee_per_item = 0
         transaction_fee = transaction_fee_per_item * sell_amount
-        item_price = user_item.item.get_current_price()
+        item_price = item.get_current_price()
         sell_price = sell_amount * item_price
         #if g.current_user.balance + sell_price < transaction_fee: abort(400, f"You cannot afford the transaction fee for this sale. The fee is [{transaction_fee}] and you will have [{g.current_user.balance + sell_price}] after sale.")
 
         try:
             g.current_user.balance += sell_price - transaction_fee
-            user_item.item.amount += sell_amount
+            item.amount += sell_amount
             user_item.amount -= sell_amount
             if user_item.amount == 0:
                 db.session.delete(user_item)
